Separate api logic from views
