+++ src/views/authors/AuthorDetails.js
