+++ src/views/insights/InsightsHome.js
