+++ src/views/insights/SalesInsights.js
