+++ src/views/shared/helpers/capitalize.js
