+++ src/views/shared/helpers/linkify.js
